Moved Viz to VisualizationHelperPrograms, which was renamed from ExtractingBuffersOfEverything.
